msm: kgsl: Use kzalloc to allocate scatterlists of 1 page or less
The majority of the scatterlist allocations used in KGSL are under 1
page (1 page of struct scatterlist is approximately 1024 entries
equalling 4MB of allocated buffer).  In these cases using vmalloc
for the sglist is undesirable and slow.  Add functions to check the
size of the allocation and favor kzalloc for 1 page allocations and
vmalloc for larger lists.

Signed-off-by: Jordan Crouse <jcrouse@codeaurora.org>
(cherry-picked from commit a652a07b0d719c4925ca45c018432888dd2388a4)

Conflicts:
	drivers/gpu/msm/kgsl_sharedmem.h

Change-Id: I6b676b4d72f20b43a4cbf3a51099aaa2e8aa8b66
Signed-off-by: Jeremy Gebben <jgebben@quicinc.com>